implement log and find correct filepaths. testing still needed
need to find a place for the log. Unsure where it should be placed.
likely in a custom folder in the documents section. see if plausible

need to find how to use files properly in package. its not right right
now. Serialization is incorrect, however with the log form,it might not
be needed. 

In that case, delete errortransfer and make a subclass for the transfer
of error codes in the ErrorID class. simple strings. 

also worked on InstanceButton implementation and their action when
pressed